requirements.txt
alpaca-py
pandas
numpy
torch
pytorch-forecasting
ta-lib
python-dotenv